20251113 wangpengfei
进销存、教务数据库切换。门店点单页面优化、商品配置字段更新

+++ miniprogram/pages/Location/Location.js
@@ -41,21 +41,46 @@ Page({
     var finduser = app.globalData.finduser
     var gongsi = app.globalData.gongsi
     var _openid = wx.getStorageSync('openid').openid;
-    wx.cloud.callFunction({
-      name: "sqlConnection",
-      data: {
-        sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
-      },
-      success(res) {
-        console.log("成功", res)
-        that.setData({
-          all: res.result
-        })
-      },
-      fail(res) {
-        console.log("失败", res)
-      }
-    });
+
+    if(app.globalData.shujuku==0){
+
+      wx.cloud.callFunction({
+        name: "sqlConnection",
+        data: {
+          sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
+        },
+        success(res) {
+          console.log("成功", res)
+          that.setData({
+            all: res.result
+          })
+        },
+        fail(res) {
+          console.log("失败", res)
+        }
+      });
+
+    }else if(app.globalData.shujuku == 1){
+
+      wx.cloud.callFunction({
+        name: "sqlServer_117",
+        data: {
+          querysql: "select * from yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql where gongsi = '" + gongsi + "'"
+        },
+        success(res) {
+          console.log("成功", res)
+          that.setData({
+            all: res.result.recordset
+          })
+        },
+        fail(res) {
+          console.log("失败", res)
+        }
+      });
+      
+    }
+
+    
   },
 
   xixi: function(e) {
@@ -66,43 +91,94 @@ Page({
       var app = getApp();
       var finduser = app.globalData.finduser
       var gongsi = app.globalData.gongsi
-      wx.cloud.callFunction({
-        name: "sqlConnection",
-        data: {
-          sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
-        },
-        success(res) {
-          that.setData({
-            all: res.result
-          })
-          console.log(that.data.szzhi)
-        },
-        fail(res) {
-          console.log("失败", res)
 
-        }
-      });
+      if(app.globalData.shujuku==0){
+
+        wx.cloud.callFunction({
+          name: "sqlConnection",
+          data: {
+            sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
+          },
+          success(res) {
+            that.setData({
+              all: res.result
+            })
+            console.log(that.data.szzhi)
+          },
+          fail(res) {
+            console.log("失败", res)
+  
+          }
+        });
+
+      }else if(app.globalData.shujuku == 1){
+
+        wx.cloud.callFunction({
+          name: "sqlServer_117",
+          data: {
+            query: "select * from yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql where gongsi = '" + gongsi + "'"
+          },
+          success(res) {
+            that.setData({
+              all: res.result.recordset
+            })
+            console.log(that.data.szzhi)
+          },
+          fail(res) {
+            console.log("失败", res)
+  
+          }
+        });
+        
+      }
+
+    
     } else {
       var that = this
       const db = wx.cloud.database()
       var app = getApp();
       var finduser = app.globalData.finduser
       var gongsi = app.globalData.gongsi
-      wx.cloud.callFunction({
-        name: "sqlConnection",
-        data: {
-          sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "' and beizhu like '%" + e.detail.value + "%'"
-        },
-        success(res) {
-          that.setData({
-            all: res.result
-          })
-          console.log(that.data.all)
-        },
-        fail(res) {
-          console.log("失败", res)
-        }
-      });
+
+      if(app.globalData.shujuku==0){
+
+        wx.cloud.callFunction({
+          name: "sqlConnection",
+          data: {
+            sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "' and beizhu like '%" + e.detail.value + "%'"
+          },
+          success(res) {
+            that.setData({
+              all: res.result
+            })
+            console.log(that.data.all)
+          },
+          fail(res) {
+            console.log("失败", res)
+          }
+        });
+
+      }else if(app.globalData.shujuku == 1){
+
+        wx.cloud.callFunction({
+          name: "sqlServer_117",
+          data: {
+            query: "select * from yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql where gongsi = '" + gongsi + "' and beizhu like '%" + e.detail.value + "%'"
+          },
+          success(res) {
+            that.setData({
+              all:res.result.recordset
+            })
+            console.log(that.data.all)
+          },
+          fail(res) {
+            console.log("失败", res)
+          }
+        });
+        
+      }
+
+
     }
   },
 
@@ -122,22 +198,48 @@ Page({
     var gongsi = app.globalData.gongsi
 
     var _openid = wx.getStorageSync('openid').openid;
-    wx.cloud.callFunction({
-      name: "sqlConnection",
-      data: {
-        sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
-      },
-      success(res) {
-        console.log("成功", res)
-        that.setData({
-          all: res.result
-        })
-      },
-      fail(res) {
-        console.log("失败", res)
 
-      }
-    });
+    if(app.globalData.shujuku==0){
+
+      wx.cloud.callFunction({
+        name: "sqlConnection",
+        data: {
+          sql: "select * from yh_jinxiaocun_jinhuofang where gongsi = '" + gongsi + "'"
+        },
+        success(res) {
+          console.log("成功", res)
+          that.setData({
+            all: res.result
+          })
+        },
+        fail(res) {
+          console.log("失败", res)
+  
+        }
+      });
+
+    }else if(app.globalData.shujuku == 1){
+
+      wx.cloud.callFunction({
+        name: "sqlServer_117",
+        data: {
+          query: "select * from yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql where gongsi = '" + gongsi + "'"
+        },
+        success(res) {
+          console.log("成功", res)
+          that.setData({
+            all: res.result.recordset
+          })
+        },
+        fail(res) {
+          console.log("失败", res)
+  
+        }
+      });
+      
+    }
+
+
   },
 
   /**
@@ -208,23 +310,49 @@ Page({
       content: '是否删除？',
       success: function(res) {
         if (res.confirm) {
-          wx.cloud.callFunction({
-            name: "sqlConnection",
-            data: {
-              sql: "delete from yh_jinxiaocun_jinhuofang where _id = '" + id + "'"
-            },
-            success(res) {
-              console.log("成功", res)
-              that.onShow();
-              // that.setData({
-              //     all: res.data,
-              //   })
-              // szZhi = 
-            },
-            fail(res) {
-              console.log("失败", res)
-            }
-          });
+
+          if(app.globalData.shujuku==0){
+
+            wx.cloud.callFunction({
+              name: "sqlConnection",
+              data: {
+                sql: "delete from yh_jinxiaocun_jinhuofang where _id = '" + id + "'"
+              },
+              success(res) {
+                console.log("成功", res)
+                that.onShow();
+                // that.setData({
+                //     all: res.data,
+                //   })
+                // szZhi = 
+              },
+              fail(res) {
+                console.log("失败", res)
+              }
+            });
+
+          }else if(app.globalData.shujuku == 1){
+
+            wx.cloud.callFunction({
+              name: "sqlServer_117",
+              data: {
+                sql: "delete fromyh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql where _id = '" + id + "'"
+              },
+              success(res) {
+                console.log("成功", res)
+                that.onShow();
+                // that.setData({
+                //     all: res.data,
+                //   })
+                // szZhi = 
+              },
+              fail(res) {
+                console.log("失败", res)
+              }
+            });
+            
+          }
+
         } else if (res.cancel) {
           return false;
         }
@@ -266,37 +394,78 @@ Page({
       const db = wx.cloud.database()
       var finduser = app.globalData.finduser
       var gongsi = app.globalData.gongsi
-      wx.cloud.callFunction({
-        name: "sqlConnection",
-        data: {
-          sql: "insert Yh_JinXiaoCun_jinhuofang (finduser,gongsi,beizhu,lianxifangshi,lianxidizhi) VALUES('" + finduser + "','" + gongsi + "','" + beizhu + "','" + lianxifangshi + "','" + lianxidizhi + "')"
-        },
-        success(res) {
-          console.log("成功", res)
-          wx.showToast({
-            title: '添加成功',
-          })
-          that.onShow()
-          wx.stopPullDownRefresh()
-        },
-        fail(res) {
-          console.log("失败", res)
 
-        }
-        //     db.collection("Yh_JinXiaoCun_jinhuofang").add({
-        //       data:{
-        //         gongsi: gongsi,
-        //         finduser: finduser,
-        //         beizhu: beizhu,
-        //         lianxifangshi: lianxifangshi,
-        //         lianxidizhi: lianxidizhi
-        //       },
-        //       success (res){
-        //        wx.showToast({
-        //        title: '添加成功',
-        // })
-        // }
-      })
+      if(app.globalData.shujuku==0){
+
+        wx.cloud.callFunction({
+          name: "sqlConnection",
+          data: {
+            sql: "insert Yh_JinXiaoCun_jinhuofang (finduser,gongsi,beizhu,lianxifangshi,lianxidizhi) VALUES('" + finduser + "','" + gongsi + "','" + beizhu + "','" + lianxifangshi + "','" + lianxidizhi + "')"
+          },
+          success(res) {
+            console.log("成功", res)
+            wx.showToast({
+              title: '添加成功',
+            })
+            that.onShow()
+            wx.stopPullDownRefresh()
+          },
+          fail(res) {
+            console.log("失败", res)
+  
+          }
+          //     db.collection("Yh_JinXiaoCun_jinhuofang").add({
+          //       data:{
+          //         gongsi: gongsi,
+          //         finduser: finduser,
+          //         beizhu: beizhu,
+          //         lianxifangshi: lianxifangshi,
+          //         lianxidizhi: lianxidizhi
+          //       },
+          //       success (res){
+          //        wx.showToast({
+          //        title: '添加成功',
+          // })
+          // }
+        })
+
+      }else if(app.globalData.shujuku == 1){
+
+        wx.cloud.callFunction({
+          name: "sqlServer_117",
+          data: {
+            query: "insert into yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql (finduser,gongsi,beizhu,lianxifangshi,lianxidizhi) VALUES('" + finduser + "','" + gongsi + "','" + beizhu + "','" + lianxifangshi + "','" + lianxidizhi + "')"
+          },
+          success(res) {
+            console.log("成功", res)
+            wx.showToast({
+              title: '添加成功',
+            })
+            that.onShow()
+            wx.stopPullDownRefresh()
+          },
+          fail(res) {
+            console.log("失败", res)
+  
+          }
+          //     db.collection("Yh_JinXiaoCun_jinhuofang").add({
+          //       data:{
+          //         gongsi: gongsi,
+          //         finduser: finduser,
+          //         beizhu: beizhu,
+          //         lianxifangshi: lianxifangshi,
+          //         lianxidizhi: lianxidizhi
+          //       },
+          //       success (res){
+          //        wx.showToast({
+          //        title: '添加成功',
+          // })
+          // }
+        })
+        
+      }
+
+
 
     }
     that.setData({
@@ -333,31 +502,65 @@ Page({
     var _this = this;
     let updIndex = _this.data.updIndex
 
-    wx.cloud.callFunction({
-      name: "sqlConnection",
-      data: {
-        sql: "update Yh_JinXiaoCun_jinhuofang set beizhu = '"+e.detail.beizhu+"',lianxifangshi= '"+e.detail.lianxifangshi+"',lianxidizhi= '"+e.detail.lianxidizhi+"' where _id = '" + _this.data.all[updIndex]._id + "'"
-      },
-      success(res) {
-        if(res.errMsg == 'cloud.callFunction:ok'){
-          _this.setData({
-            updIndex: -1
-          }, function(){
-            _this.onShow()
-            wx.showToast({
-              title: '修改成功',
+    if(app.globalData.shujuku==0){
+
+      wx.cloud.callFunction({
+        name: "sqlConnection",
+        data: {
+          sql: "update Yh_JinXiaoCun_jinhuofang set beizhu = '"+e.detail.beizhu+"',lianxifangshi= '"+e.detail.lianxifangshi+"',lianxidizhi= '"+e.detail.lianxidizhi+"' where _id = '" + _this.data.all[updIndex]._id + "'"
+        },
+        success(res) {
+          if(res.errMsg == 'cloud.callFunction:ok'){
+            _this.setData({
+              updIndex: -1
+            }, function(){
+              _this.onShow()
+              wx.showToast({
+                title: '修改成功',
+              })
             })
+          }
+        },
+        fail(res) {
+          wx.showToast({
+            title: '修改失败',
+            icon: 'none',
+            mask: true
           })
         }
-      },
-      fail(res) {
-        wx.showToast({
-          title: '修改失败',
-          icon: 'none',
-          mask: true
-        })
-      }
-    });
+      });
+
+    }else if(app.globalData.shujuku == 1){
+
+      wx.cloud.callFunction({
+        name: "sqlServer_117",
+        data: {
+          query: "update yh_jinxiaocun_excel.dbo.yh_jinxiaocun_jinhuofang_mssql set beizhu = '"+e.detail.beizhu+"',lianxifangshi= '"+e.detail.lianxifangshi+"',lianxidizhi= '"+e.detail.lianxidizhi+"' where _id = '" + _this.data.all[updIndex]._id + "'"
+        },
+        success(res) {
+          if(res.errMsg == 'cloud.callFunction:ok'){
+            _this.setData({
+              updIndex: -1
+            }, function(){
+              _this.onShow()
+              wx.showToast({
+                title: '修改成功',
+              })
+            })
+          }
+        },
+        fail(res) {
+          wx.showToast({
+            title: '修改失败',
+            icon: 'none',
+            mask: true
+          })
+        }
+      });
+      
+    }
+
+    
   },
 
   back: function(){
